+++ cmd/study/list.go
@@ -9,15 +9,14 @@ import (
 	"github.com/benmatselby/prolificli/client"
 	"github.com/benmatselby/prolificli/model"
 	"github.com/benmatselby/prolificli/ui/study"
-	"github.com/charmbracelet/bubbles/list"
-	tea "github.com/charmbracelet/bubbletea"
 	"github.com/spf13/cobra"
 )
 
 // ListOptions is the options for the listing studies command.
 type ListOptions struct {
-	Args   []string
-	Status string
+	Args           []string
+	Status         string
+	NonInteractive bool
 }
 
 // NewListCommand creates a new `study list` command to give you details about
@@ -31,7 +30,18 @@ func NewListCommand(commandName string, client client.API, w io.Writer) *cobra.C
 		Run: func(cmd *cobra.Command, args []string) {
 			opts.Args = args
 
-			err := renderList(client, opts, w)
+			renderer := study.ListRenderer{}
+
+			if opts.NonInteractive {
+				renderer.SetStrategy(&study.NonInteractiveRenderer{})
+			} else {
+				renderer.SetStrategy(&study.InteractiveRenderer{})
+			}
+
+			err := renderer.Render(client, study.ListUsedOptions{
+				Status: opts.Status, NonInteractive: opts.NonInteractive,
+			}, w)
+
 			if err != nil {
 				fmt.Printf("Error: %s", strings.ReplaceAll(err.Error(), "\n", ""))
 				os.Exit(1)
@@ -41,34 +51,7 @@ func NewListCommand(commandName string, client client.API, w io.Writer) *cobra.C
 
 	flags := cmd.Flags()
 	flags.StringVarP(&opts.Status, "status", "s", model.StatusAll, fmt.Sprintf("The status you want to filter on: %s.", strings.Join(model.StudyListStatus, ", ")))
+	flags.BoolVarP(&opts.NonInteractive, "non-interactive", "n", false, "Render the list details straight to the terminal.")
 
 	return cmd
-}
-
-func renderList(client client.API, opts ListOptions, w io.Writer) error {
-	studies, err := client.GetStudies(opts.Status)
-	if err != nil {
-		return err
-	}
-
-	var items []list.Item
-	var studyMap = make(map[string]model.Study)
-
-	for _, study := range studies.Results {
-		items = append(items, study)
-		studyMap[study.ID] = study
-	}
-
-	lv := study.ListView{
-		List:    list.New(items, list.NewDefaultDelegate(), 0, 0),
-		Studies: studyMap,
-		Client:  client,
-	}
-	lv.List.Title = "My studies"
-
-	if err := tea.NewProgram(lv).Start(); err != nil {
-		return fmt.Errorf("cannot render studies: %s", err)
-	}
-
-	return nil
 }